Back end

+++ back-end/src/modules/customer/customer.service.ts
@@ -1,4 +1,8 @@
-import { Injectable } from '@nestjs/common';
+import {
+  ConflictException,
+  Injectable,
+  NotFoundException,
+} from '@nestjs/common';
 import { CreateCustomerDto } from './dto/create-customer.dto';
 import { UpdateCustomerDto } from './dto/update-customer.dto';
 import { CustomerRepository } from './repositories/customer.repository';
@@ -8,18 +12,35 @@ export class CustomerService {
   constructor(protected customerRepository: CustomerRepository) {}
 
   async create(createCustomerDto: CreateCustomerDto) {
+    const findCustomer = await this.customerRepository.findByMail(
+      createCustomerDto.email,
+    );
+    if (findCustomer) {
+      throw new ConflictException('Cliente já cadastrado');
+    }
     const customer = await this.customerRepository.create(createCustomerDto);
 
     return customer;
   }
 
   async findAll() {
-    const customer = await this.customerRepository.findAll();
-    return customer;
+    const customers = await this.customerRepository.findAll();
+    return customers;
   }
 
   async findOne(id: string) {
     const customer = await this.customerRepository.findOne(id);
+    if (!customer) {
+      throw new NotFoundException('Usuário não encontrado');
+    }
+    return customer;
+  }
+
+  async findByMail(email: string) {
+    const customer = await this.customerRepository.findByMail(email);
+    // if (!customer) {
+    //   throw new NotFoundException('Usuário não encontrado');
+    // }
     return customer;
   }
 
@@ -28,10 +49,17 @@ export class CustomerService {
       id,
       updateCustomerDto,
     );
+    if (!customer) {
+      throw new NotFoundException('Usuário não encontrado');
+    }
     return customer;
   }
 
   async remove(id: string) {
+    const customer = await this.customerRepository.findOne(id);
+    if (!customer) {
+      throw new NotFoundException('Usuário não encontrado');
+    }
     await this.customerRepository.delete(id);
     return;
   }
